feat: footer 로고 추가
