--- src/components/layouts/Footer.tsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-const Footer = () => {
-  return <SFooter>푸터영역</SFooter>;
-};
-
-export default Footer;
-
-const SFooter = styled.footer`
-  padding-bottom: ${({ theme }) => theme.layouts.mobileNavHeight};
-  height: ${({ theme }) => theme.layouts.footerHeight};
-  background-color: ${({ theme }) => theme.colors.gray100};
-`;
